restaurant/requirements.txt
Django==4.1.1
psycopg2==2.9.3
djangorestframework==3.13.1
pytest==7.3.1
jwt